feat(components): pageLayout component updated
sider moved to separate component
theme switcher moved to sider
student nav moved to page header

META DATA: issue39
