+++ package.json
@@ -20,7 +20,8 @@
     "history": "^5.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-uuid": "^2.0.0"
+    "react-uuid": "^2.0.0",
+    "vite-plugin-svgr": "^2.4.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.3.0",
