+++ src/pages/student/page.tsx
@@ -1,6 +1,6 @@
 import { PageLayout } from "../../shared/ui";
 
-const nav = ["home", "student", "schedule"];
+const nav = ["home", "schedule"];
 
 export const StudentPage = () => (
   <PageLayout title="Личный кабинет" nav={nav} />
