+++ src/pages/student/settings/page.tsx
@@ -1,6 +1,6 @@
 import { PageLayout } from "../../../shared/ui";
 
-const nav = ["home", "student", "schedule"];
+const nav = ["home", "schedule"];
 
 export const StudentSettingsPage = () => (
   <PageLayout title="Настройки" nav={nav} />
