+++ src/shared/ui/Icons/index.ts
@@ -1,0 +1,2 @@
+export { ReactComponent as IconMoon } from "./lib/moon.svg";
+export { ReactComponent as IconSun } from "./lib/sun.svg";
