+++ src/shared/ui/PageLayout/PageHeader/PageHeader.tsx
@@ -1,5 +1,7 @@
-import { Layout } from "antd";
-import { ThemeSwitcher } from "../../../../features/theme-switcher";
+import { UserOutlined, SettingOutlined } from "@ant-design/icons";
+import { Layout, Menu, MenuProps } from "antd";
+import { Link } from "atomic-router-react";
+import { routes } from "../../../lib/atomic-router/route";
 import styles from "./PageHeader.module.css";
 
 const { Header } = Layout;
@@ -8,9 +10,49 @@ type Props = {
   title: string;
 };
 
+type MenuItem = Required<MenuProps>["items"][number];
+
+function getItem(
+  label: React.ReactNode,
+  key: React.Key,
+  icon?: React.ReactNode,
+  children?: MenuItem[]
+): MenuItem {
+  return {
+    key,
+    icon,
+    children,
+    label,
+  } as MenuItem;
+}
+
+const items = [
+  getItem("Личный кабинет", "studentTop", <UserOutlined />, [
+    getItem(
+      <Link to={routes.student}>
+        <u>Личный кабинет</u>
+      </Link>,
+      "student",
+      <UserOutlined />
+    ),
+    getItem(
+      <Link to={routes.studentSettings}>
+        <u>Настройки</u>
+      </Link>,
+      "studentSettings",
+      <SettingOutlined />
+    ),
+  ]),
+];
+
 export const PageHeader = ({ title }: Props) => (
   <Header className={styles._}>
     <span className={styles.title}>{title}</span>
-    <ThemeSwitcher />
+    <Menu
+      className={styles.menu}
+      theme="dark"
+      mode="horizontal"
+      items={items}
+    />
   </Header>
 );
