+++ src/shared/ui/index.ts
@@ -1,2 +1,3 @@
 export { PageLayout } from "./PageLayout/PageLayout";
 export { PageHeader } from "./PageLayout/PageHeader/PageHeader";
+export { NavigationSider } from "./PageLayout/NavigationSider/NavigationSider";
