+++ vite.config.ts
@@ -1,7 +1,8 @@
 import { defineConfig } from 'vite'
+import svgr from 'vite-plugin-svgr'
 import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()]
+  plugins: [react(), svgr()]
 })
